+++ CHANGELOG.md
@@ -21,11 +21,12 @@
     available configuration options for the server.
   - [Work in progress] Define the `ClientManager` class to manage references to
     connected clients.
-- [Work in progress] Define the `Socket` class as the super-type of sockets.
+- [Work in progress] Define the `Socket` class as the super-type of sockets:
   - [Work in progress] Define the `HeartbeatManager` class to manage the
     heartbeat mechanism.
   - [Work in progress] Define the server/`Socket` class to represent clients
     connected to the server.
 - [Work in progress] Define the `EngineException` class:
   - [Work in progress] Define the `ServerException` class.
-  - [Work in progress] Define the `TransportException` class.
+  - [Work in progress] Define the `TransportException` class:
+    - [Work in progress] Define the `PollingTransportException` class.